requirements.txt
PyYAML~=6.0
rich~=11.1.0